Adding in three data files, provided comments on all core functionality and how it is triggered; TODO for submitting buffers to the user and some comments on the helper functions still need completed
Mostly feature complete;  Missing packing data into a buffer since it relies heavily on how we want to do choosing of measurements and what it means to "deliver a buffer to a user".

Find the first occurrence & second occurrence are finished but need some slight testing.  (Demoes choosing a field to search on)
Summary per data file we choose to write to is also complete and has been tested thoroughly.  (Demoes some of the working with the file system)
Successfully read in and store the data for ourselves and have thoroughly tested this also.

I eneded up doing simple csv since I got home later than expected.